requirements/core.txt
aiohttp
asyncpg
pyyaml